python-scripts/requirements.txt
requests
pyjwt
schedule
